Publish both generated API clients
public-api-client could not be published as it stood: its hand-written
operation layer imported runtime zod schemas from bookings, finance and
public-api, declared as devDependencies and peerDependencies, so an
install from npm would have failed at import. verify:public-surface
catches exactly that, and did.

That layer covered 23 of 138 paths and has one consumer, so it moves to
public-api-react, which is private and already depends on those packages.
What ships is the generated surface plus the typed constructor.

The key prefix table moves from core to public-api-contracts. The clients
need to classify a token; core is the kernel - container, saga, registry,
project - and publishing it so a browser can check a prefix would commit
us to a public API nobody intended. contracts is already published and is
where a public API key's shape belongs. core re-exports it, so the eleven
in-repo callers are untouched, and there is still exactly one table.

Public surface closure: 18 packages, closed.
