+++ packages/core/package.json
@@ -31,7 +31,8 @@
     "prepack": "pnpm run build"
   },
   "dependencies": {
-    "@voyant-travel/graph-contracts": "workspace:^"
+    "@voyant-travel/graph-contracts": "workspace:^",
+    "@voyant-travel/public-api-contracts": "workspace:^"
   },
   "devDependencies": {
     "@voyant-travel/voyant-typescript-config": "workspace:^",
