+++ packages/public-api-contracts/src/index.ts
@@ -1,2 +1,3 @@
+export * from "./public-api-key.js"
 export * from "./bookings.js"
 export * from "./profile.js"
