+++ packages/public-api-react/package.json
@@ -74,6 +74,9 @@
     }
   },
   "dependencies": {
+    "@voyant-travel/bookings": "workspace:^",
+    "@voyant-travel/bookings-contracts": "workspace:^",
+    "@voyant-travel/finance": "workspace:^",
     "@voyant-travel/i18n": "workspace:^",
     "@voyant-travel/public-api-contracts": "workspace:^",
     "@voyant-travel/react": "workspace:^"
