+++ packages/public-api-react/src/legacy-client/booking-session-v1.ts
