+++ packages/public-api-react/src/legacy-client/errors.ts
